% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcga_available.R
\name{tcga_available}
\alias{tcga_available}
\title{Prints available TCGA datasets}
\usage{
tcga_available()
}
\description{
Prints available TCGA cohorts
}
\examples{
tcga_available()
}
\seealso{
\code{\link{tcga_load}}
}
